# Add weaklist benchmarks