@smoogipoo 2020-01-09
One more, sorry.

It does show where some improvement needs to happen, at least. Wouldn't be surprised if that's a blocking point for our disposal - the unbinding hitting that `RemoveAllAndEnumerate` path very hard.